cloud-functions/generate-pdfs/requirements.txt
functions-framework==3.*
python-dotenv==1.*
requests==2.*
python-docx==1.*
python-pptx==1.*
google-cloud-storage==2.*
google-auth